refactor: filenames, footer inside layout
