+++ src/pages/blog-page.tsx
@@ -1,0 +1,13 @@
+const BlogPage = () => {
+  return (
+    <div className='pt-16 px-12 mt-8'>
+      <div>
+        <h2 className='text-4xl text-slate-900 font-roboto-slab dark:text-slate-50 font-bold relative after:contents[] after:w-36 after:h-[2px] after:bg-primary after:absolute after:top-1/2 -translate-y-1/2 after:right-8 after:left-52'>
+          Blog
+        </h2>
+      </div>
+    </div>
+  );
+};
+
+export default BlogPage;
